pyproject.toml
[tool.poetry]
name = "python-django"
version = "0.1.0"
description = "Python-Search Project"
authors = ["euiyounghwang <marieuig@gmail.com>"]
readme = "README.md"
packages = [{include = "python_django"}]

[tool.poetry.dependencies]
python = "^3.9"
elasticsearch = "7.13.0"
django = "^4.2.4"
djangorestframework = "^3.14.0"
django-prometheus = "^2.3.1"
markdown = "^3.5"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.7"
python-dotenv = "1.0.0"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.0"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-django = "^4.5.2"
django-redis = "^5.4.0"
redis = "^5.0.1"
retry = "^0.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
